+++ packages/toprf-secure-backup/src/index.ts
@@ -3,3 +3,10 @@ export * from './toprfSecureBackup';
 export type { RateLimitErrorData } from './errors';
 export { TOPRFError, TOPRFErrorCode } from './errors';
 export * from './utils';
+export { EncAccountDataType } from './constants';
+export type {
+  SecretDataItem,
+  UpdateSecretDataItem,
+  MetadataUpdateSecretDataItemParams,
+  MetadataBatchUpdateSecretDataItemParams,
+} from './metadata';